test/integration/pyproject.toml
# Copyright (C) 2024 Roberto Rossini <roberros@uio.no>
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "hictk_integration_suite"
version = "0.0.1"
description = "Integration test suite for hictk"
authors = [
  {name = "Roberto Rossini", email = "roberros@uio.no"}
]
requires-python = ">=3.11"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License"
]

dependencies = [
  "click>=8",
  "colorama==0.4.*",
  "cooler==0.10.*",
  "h5py!=3.12",  # this version is broken on Windows
  "hictkpy[pandas]==1.1.*",
  "immutabledict==4.*",
  "numpy",
  "pandas==2.2.*",
  "PyYAML",
  "pywin32>=308; platform_system == 'Windows'",
  "structlog==25.*",
]

optional-dependencies.test = [
  "pytest>=8.0"
]

optional-dependencies.dev = [
  "black",
  "isort",
  "hictk_integration_suite[test]",
  "pre-commit",
]

[project.scripts]
hictk_integration_suite = "hictk_integration_suite:main"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
  "ignore:(ast.Str|Attribute s|ast.NameConstant|ast.Num) is deprecated:DeprecationWarning:_pytest",
  "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\)*:DeprecationWarning",  # https://github.com/pytest-dev/pytest/issues/11528
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.hatch.envs.default.scripts]
fix = "black src test"
lint = "black --check src test"
test = "pytest test"

[tool.hatch.envs.test]
features = ["dev"]

[[tool.hatch.envs.test.matrix]]
python = ["3.11", "3.12", "3.13"]
